pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-plugin"
version = "0.1.0"
description = "CLI tool for AI-powered social media content generation and publishing"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "tweepy>=4.14",
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
    "google-auth-oauthlib>=1.1",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "httpx>=0.25",
    "pdfplumber>=0.10",
    "Pillow>=10.0",
    "tenacity>=8.2",
    "python-dateutil>=2.8",
    "better-profanity>=0.7",
    "slack-sdk>=3.23",
    "python-docx>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.11",
]

[project.scripts]
social-plugin = "social_plugin.cli:cli"

[tool.setuptools.packages.find]
include = ["social_plugin*"]
